dbd_oracle: Enable threaded mode properly
Instead of enabling threaded mode in OCIInitialize(), which is
deprecated, and then overriding it to OCI_DEFAULT in OCIEnvCreate,
don't OCIInitialize() anymore, and use OCI_THREADED mode with
OCIEnvCreate().

OCIEnvCreate() is recommended over the OCIInitialize() + OCIEnvInit()
pair anyway, so removing OCIInitialize() is safe, and OCI_THREADED
must be set for OCIEnvCreate().

This should - if all goes well - fix the threading crashes with oracle
destinations.