configure: port --enable-mysql-threadsafe from upstream 1/2
Since the beginning of time, syslog-ng connects to MySQL servers from separate
worker threads. As stated in the documentation of libmysqlclient:

> If your client programs are using threads, you must compile a
> thread-safe version of the MySQL client library with the
> `--enable-thread-safe-client' configure option. This creates a
> `libmysqlclient_r' library with which you should link your
> threaded applications. See *Note threaded-clients::.

Sadly, the libdbi version syslog-ng depends on, is linked against
`libmysqlclient.so` only, so we need to backport some patches enabling linking
against `libmysqlclient_r.so` from upstream.

This patch is equivalent to:

    revision 1.28
    date: 2010/09/21 21:12:39;  author: mhoenicka;  state: Exp;  lines: +7 -3
    Jans patch for linking the mysql driver properly to libmysqlclient

References: bugzilla #30089