freetds: fix buffer overflow
_dbd_decode_binary is the reverse method of a binary encoding algorithm
that is originated from the SQLite library.
They removed their version from SQLite 3 in 2004.

This function expects the input to be null-terminated. It has nothing to
do with raw binary fields.

Neither the buffer nor the database field was null-terminated, so it could
overwrite adjacent memory locations: buffer overflow.

In case of binary fields, _dbd_decode_binary() should not be used, the
raw result can be returned instead.

The new bzero() call is here to avoid further errors.

Signed-off-by: László Várady <laszlo.varady@balabit.com>