freetds: fix memleak in decimal->string conversion
During the decimal->string covnersion, we allocate a string,
which does not get freed in `_dbd_free_row()`, because only
DBI_TYPE_STRING and DBI_TYPE_BINARY type query value is
freed there. So I changed the value type to DBI_TYPE_STRING.

DBI_TYPE_NUMERIC_AS_STRING became unnecessary when we use
DBI_TYPE_STRING instead of it. The commit, which introduces
it is reverted in libdbi.

Signed-off-by: Attila Szakacs <attila.szakacs@balabit.com>