sbuild: look for oracle in lib instead of oracle/lib
tgz2build already does it this way. This means installdeps already
installs the oracle libraries under lib. So that this to work, libdbi
is also hacked: sbuild searches under oracle/lib, tgz2build searches
under lib. If someone wants to develop either one, both libraries need
to be compiled locally, which is something easy to miss.

In case of libdi, the location is specified by configure
options. There is no such option in case of odbc. For odbc, sbuild
either need to install under lib too, or we need to add oracle/lib to
LD_LIBRARY_PATH.

Making sbuild work the same as tgz2build is cleaner, and solves the
odbc problem.

Signed-off-by: Antal Nemes <antal.nemes.hu@gmail.com>